+++ include/Gaffer/ShufflePlug.h
@@ -95,11 +95,10 @@ class GAFFER_API ShufflesPlug : public ValuePlug
 		bool acceptsInput( const Plug *input ) const override;
 		PlugPtr createCounterpart( const std::string &name, Direction direction ) const override;
 
-		/// Shuffles the sources into a destination container. The container type must have a std::pair value_type
-		/// and string-compatible keys (eg std::string, IECore::InternedString).
+		/// Shuffles the sources into a destination container. The container type should have a map
+		/// compatible interface with string-compatible keys (eg std::string, IECore::InternedString).
 		template<typename T>
 		T shuffle( const T &sourceContainer ) const;
-
 };
 
 } // namespace Gaffer